Allow rebalancing in setContributionSettings()
Some edits to the contribution settings will force us to rebalance the entire
pool. However, there are a class of edits to contribution settings where
the pool already maintains the invariants for the new limits. In those cases
we do not rebalance the pool as an optimization.

This change allows the caller to supply a list of addresses to rebalance in
case there are contributors are able to increase their contribution as a result
of limit changes.